Clean up toolkit install script
