+++ scripts/update-toolkit-package.js
@@ -1,4 +1,3 @@
-const {color} = require('./helpers');
 const {exec} = require('child_process');
 const fs = require('fs');
 const process = require('process');
@@ -7,22 +6,20 @@ const util = require('util');
 const execShellCommand = util.promisify(exec);
 
 /**
- * This script updates the toolkit package to the latest version in
+ * This script installs the latest version of the toolkit package in
  * every sample extension.
  */
 async function main() {
 	console.log('Updating toolkit package to latest version...');
   console.log();
 
-  await updateToolkitPackage('default');
-  await updateToolkitPackage('react', 'webview-ui');
+  await installLatestToolkitPackage('default');
+  await installLatestToolkitPackage('react', 'webview-ui');
 
-	// Print success and next steps messages
-	console.log();
 	console.log(color(['bold', 'green'], 'All extension samples updated to the latest version of the toolkit!'));
 }
 
-async function updateToolkitPackage(rootDir, installDir = '') {
+async function installLatestToolkitPackage(rootDir, installDir = '') {
   try {
     const dirs = getSubDirectories(rootDir);
     for (const dir of dirs) {
@@ -42,6 +39,23 @@ function getSubDirectories(source) {
   const dirContent = fs.readdirSync(source, { withFileTypes: true });
   const subDirectoryList = dirContent.filter(content => content.isDirectory()).map(dir => dir.name);
   return subDirectoryList;
+}
+
+const colors = {
+	reset: '\x1b[0m',
+	bold: '\x1b[1m',
+	dim: '\x1b[2m',
+	red: '\x1b[31m',
+	green: '\x1b[32m',
+	cyan: '\x1b[36m',
+};
+
+function color(opts, text) {
+	let colorString = '';
+	for (const opt of opts) {
+		colorString += colors[opt];
+	}
+	return `${colorString}${text}${colors.reset}`;
 }
 
 main();
